+++ reliability/Fitters.py
@@ -28,7 +28,7 @@ import pandas as pd
 from scipy.optimize import minimize, curve_fit
 import scipy.stats as ss
 import warnings
-from reliability.Distributions import Weibull_Distribution, Gamma_Distribution, Beta_Distribution, Exponential_Distribution, Normal_Distribution, Lognormal_Distribution, Mixture_Model, Competing_Risks_Model
+from reliability.Distributions import Weibull_Distribution, Gamma_Distribution, Beta_Distribution, Exponential_Distribution, Normal_Distribution, Lognormal_Distribution, Loglogistic_Distribution, Mixture_Model, Competing_Risks_Model
 from reliability.Nonparametric import KaplanMeier
 from reliability.Probability_plotting import plotting_positions
 from reliability.Utils import round_to_decimals
@@ -120,14 +120,14 @@ class Fit_Everything:
         if type(right_censored) != np.ndarray:
             raise TypeError('right_censored must be a list or array of right censored failure data')
 
-        #remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
+        # remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
         f0 = np.array(failures)
         failures = f0[f0 != 0]
-        if len(failures)!=len(f0):
+        if len(failures) != len(f0):
             print('WARNING: failures contained zeros. These have been removed to enable fitting.')
         rc0 = np.array(right_censored)
         right_censored = rc0[rc0 != 0]
-        if len(right_censored)!=len(rc0):
+        if len(right_censored) != len(rc0):
             print('WARNING: right_censored contained zeros. These have been removed to enable fitting.')
 
         self.failures = failures
@@ -310,11 +310,11 @@ class Fit_Everything:
         bin_width = 2 * iqr * len(X) ** -(1 / 3)  # Freedman–Diaconis rule ==> https://en.wikipedia.org/wiki/Freedman%E2%80%93Diaconis_rule
         num_bins = int(np.ceil((max(X) - min(X)) / bin_width))
         hist, bins = np.histogram(X, bins=num_bins, density=True)
-        upper_lim = max(hist)*1.5
+        upper_lim = max(hist) * 1.5
         hist_cumulative = np.cumsum(hist) / sum(hist)
         width = np.diff(bins)
         center = (bins[:-1] + bins[1:]) / 2
-        plt.bar(center, hist * self._frac_fail, align='center', width=width, color='lightgrey', edgecolor='k',linewidth=0.5)
+        plt.bar(center, hist * self._frac_fail, align='center', width=width, color='lightgrey', edgecolor='k', linewidth=0.5)
 
         Weibull_Distribution(alpha=self.Weibull_2P_alpha, beta=self.Weibull_2P_beta).PDF(xvals=xvals, label=r'Weibull ($\alpha , \beta$)')
         Weibull_Distribution(alpha=self.Weibull_3P_alpha, beta=self.Weibull_3P_beta, gamma=self.Weibull_3P_gamma).PDF(xvals=xvals, label=r'Weibull ($\alpha , \beta , \gamma$)')
@@ -329,14 +329,14 @@ class Fit_Everything:
             Beta_Distribution(alpha=self.Beta_2P_alpha, beta=self.Beta_2P_beta).PDF(xvals=xvals, label=r'Beta ($\alpha , \beta$)')
         plt.legend()
         plt.xlim(xmin, xmax)
-        plt.ylim(0,upper_lim)
+        plt.ylim(0, upper_lim)
         plt.title('Probability Density Function')
         plt.xlabel('Data')
         plt.ylabel('Probability density')
         plt.legend()
 
         plt.subplot(122)  # CDF
-        plt.bar(center, hist_cumulative * self._frac_fail, align='center', width=width, color='lightgrey', edgecolor='k',linewidth=0.5)
+        plt.bar(center, hist_cumulative * self._frac_fail, align='center', width=width, color='lightgrey', edgecolor='k', linewidth=0.5)
         Weibull_Distribution(alpha=self.Weibull_2P_alpha, beta=self.Weibull_2P_beta).CDF(xvals=xvals, label=r'Weibull ($\alpha , \beta$)')
         Weibull_Distribution(alpha=self.Weibull_3P_alpha, beta=self.Weibull_3P_beta, gamma=self.Weibull_3P_gamma).CDF(xvals=xvals, label=r'Weibull ($\alpha , \beta , \gamma$)')
         Gamma_Distribution(alpha=self.Gamma_2P_alpha, beta=self.Gamma_2P_beta).CDF(xvals=xvals, label=r'Gamma ($\alpha , \beta$)')
@@ -653,14 +653,14 @@ class Fit_Weibull_2P:
         if type(right_censored) != np.ndarray:
             raise TypeError('right_censored must be a list or array of right censored failure data')
 
-        #remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
+        # remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
         rc0 = right_censored
         f0 = failures
         right_censored = rc0[rc0 != 0]
         failures = f0[f0 != 0]
-        if len(failures)!=len(f0):
+        if len(failures) != len(f0):
             print('WARNING: failures contained zeros. These have been removed to enable fitting.')
-        if len(right_censored)!=len(rc0):
+        if len(right_censored) != len(rc0):
             print('WARNING: right_censored contained zeros. These have been removed to enable fitting.')
 
         all_data = np.hstack([failures, right_censored])
@@ -756,7 +756,7 @@ class Fit_Weibull_2P:
                     self.beta = force_beta
 
         params = [self.alpha, self.beta]
-        k = len(params)
+        k = len(params)  ### CHECK why are k and n obtained again. Shouldn't k remain 1 if force_beta is used?
         n = len(all_data)
         LL2 = 2 * Fit_Weibull_2P.LL(params, failures, right_censored)
         self.loglik2 = LL2
@@ -931,13 +931,13 @@ class Fit_Weibull_2P_grouped:
         if optimizer not in ['L-BFGS-B', 'TNC']:
             raise ValueError('optimizer must be either "L-BFGS-B" OR "TNC". Default is "L-BFGS-B".')
 
-        if min(dataframe.time.values)<0:
-            raise ValueError('All failure and censoring times must be greater than zero.') #raise an error is there are values below 0
+        if min(dataframe.time.values) < 0:
+            raise ValueError('All failure and censoring times must be greater than zero.')  # raise an error is there are values below 0
 
         dataframe0 = dataframe
         dataframe = dataframe0[dataframe0['time'] > 0]
         if len(dataframe0.time.values) != len(dataframe.time.values):
-            print('WARNING: dataframe contained zeros. These have been removed to enable fitting.') #automatically filter out zeros and print warning if zeros have been removed
+            print('WARNING: dataframe contained zeros. These have been removed to enable fitting.')  # automatically filter out zeros and print warning if zeros have been removed
 
         # unpack the dataframe
         failures_df = dataframe[dataframe['category'] == 'F']
@@ -1213,14 +1213,14 @@ class Fit_Weibull_3P:
         if optimizer not in ['L-BFGS-B', 'TNC']:
             raise ValueError('optimizer must be either "L-BFGS-B" OR "TNC". Default is "L-BFGS-B".')
 
-        #remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
+        # remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
         rc0 = right_censored
         f0 = failures
         right_censored = rc0[rc0 != 0]
         failures = f0[f0 != 0]
-        if len(failures)!=len(f0):
+        if len(failures) != len(f0):
             print('WARNING: failures contained zeros. These have been removed to enable fitting.')
-        if len(right_censored)!=len(rc0):
+        if len(right_censored) != len(rc0):
             print('WARNING: right_censored contained zeros. These have been removed to enable fitting.')
 
         # Weibull_3P INITIAL GUESS SECTION
@@ -1473,14 +1473,14 @@ class Fit_Weibull_Mixture:
         if type(right_censored) != np.ndarray:
             raise TypeError('right_censored must be a list or array of right censored failure data')
 
-        #remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
+        # remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
         rc0 = right_censored
         f0 = failures
         right_censored = rc0[rc0 != 0]
         failures = f0[f0 != 0]
-        if len(failures)!=len(f0):
+        if len(failures) != len(f0):
             print('WARNING: failures contained zeros. These have been removed to enable fitting.')
-        if len(right_censored)!=len(rc0):
+        if len(right_censored) != len(rc0):
             print('WARNING: right_censored contained zeros. These have been removed to enable fitting.')
 
         all_data = np.hstack([failures, right_censored])
@@ -1691,14 +1691,14 @@ class Fit_Weibull_CR:
         if type(right_censored) != np.ndarray:
             raise TypeError('right_censored must be a list or array of right censored failure data')
 
-        #remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
+        # remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
         rc0 = right_censored
         f0 = failures
         right_censored = rc0[rc0 != 0]
         failures = f0[f0 != 0]
-        if len(failures)!=len(f0):
+        if len(failures) != len(f0):
             print('WARNING: failures contained zeros. These have been removed to enable fitting.')
-        if len(right_censored)!=len(rc0):
+        if len(right_censored) != len(rc0):
             print('WARNING: right_censored contained zeros. These have been removed to enable fitting.')
 
         all_data = np.hstack([failures, right_censored])
@@ -1887,14 +1887,14 @@ class Fit_Expon_1P:
         if type(right_censored) != np.ndarray:
             raise TypeError('right_censored must be a list or array of right censored failure data')
 
-        #remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
+        # remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
         rc0 = right_censored
         f0 = failures
         right_censored = rc0[rc0 != 0]
         failures = f0[f0 != 0]
-        if len(failures)!=len(f0):
+        if len(failures) != len(f0):
             print('WARNING: failures contained zeros. These have been removed to enable fitting.')
-        if len(right_censored)!=len(rc0):
+        if len(right_censored) != len(rc0):
             print('WARNING: right_censored contained zeros. These have been removed to enable fitting.')
 
         all_data = np.hstack([failures, right_censored])
@@ -2045,14 +2045,14 @@ class Fit_Expon_2P:
         if type(right_censored) != np.ndarray:
             raise TypeError('right_censored must be a list or array of right censored failure data')
 
-        #remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
+        # remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
         rc0 = right_censored
         f0 = failures
         right_censored = rc0[rc0 != 0]
         failures = f0[f0 != 0]
-        if len(failures)!=len(f0):
+        if len(failures) != len(f0):
             print('WARNING: failures contained zeros. These have been removed to enable fitting.')
-        if len(right_censored)!=len(rc0):
+        if len(right_censored) != len(rc0):
             print('WARNING: right_censored contained zeros. These have been removed to enable fitting.')
 
         all_data = np.hstack([failures, right_censored])
@@ -2422,14 +2422,14 @@ class Fit_Lognormal_2P:
         if type(right_censored) != np.ndarray:
             raise TypeError('right_censored must be a list or array of right censored failure data')
 
-        #remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
+        # remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
         rc0 = right_censored
         f0 = failures
         right_censored = rc0[rc0 != 0]
         failures = f0[f0 != 0]
-        if len(failures)!=len(f0):
+        if len(failures) != len(f0):
             print('WARNING: failures contained zeros. These have been removed to enable fitting.')
-        if len(right_censored)!=len(rc0):
+        if len(right_censored) != len(rc0):
             print('WARNING: right_censored contained zeros. These have been removed to enable fitting.')
 
         self.gamma = 0
@@ -2609,14 +2609,14 @@ class Fit_Lognormal_3P:
         if type(right_censored) != np.ndarray:
             raise TypeError('right_censored must be a list or array of right censored failure data')
 
-        #remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
+        # remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
         rc0 = right_censored
         f0 = failures
         right_censored = rc0[rc0 != 0]
         failures = f0[f0 != 0]
-        if len(failures)!=len(f0):
+        if len(failures) != len(f0):
             print('WARNING: failures contained zeros. These have been removed to enable fitting.')
-        if len(right_censored)!=len(rc0):
+        if len(right_censored) != len(rc0):
             print('WARNING: right_censored contained zeros. These have been removed to enable fitting.')
 
         all_data = np.hstack([failures, right_censored])
@@ -2770,6 +2770,7 @@ class Fit_Lognormal_3P:
         LL_rc += Fit_Lognormal_3P.logR(T_rc, params[0], params[1], params[2]).sum()  # right censored times
         return -(LL_f + LL_rc)
 
+
 class Fit_Gamma_2P:
     '''
     Fit_Gamma_2P
@@ -2825,14 +2826,14 @@ class Fit_Gamma_2P:
         if type(right_censored) != np.ndarray:
             raise TypeError('right_censored must be a list or array of right censored failure data')
 
-        #remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
+        # remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
         rc0 = right_censored
         f0 = failures
         right_censored = rc0[rc0 != 0]
         failures = f0[f0 != 0]
-        if len(failures)!=len(f0):
+        if len(failures) != len(f0):
             print('WARNING: failures contained zeros. These have been removed to enable fitting.')
-        if len(right_censored)!=len(rc0):
+        if len(right_censored) != len(rc0):
             print('WARNING: right_censored contained zeros. These have been removed to enable fitting.')
 
         all_data = np.hstack([failures, right_censored])
@@ -2998,14 +2999,14 @@ class Fit_Gamma_3P:
         if type(right_censored) != np.ndarray:
             raise TypeError('right_censored must be a list or array of right censored failure data')
 
-        #remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
+        # remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
         rc0 = right_censored
         f0 = failures
         right_censored = rc0[rc0 != 0]
         failures = f0[f0 != 0]
-        if len(failures)!=len(f0):
+        if len(failures) != len(f0):
             print('WARNING: failures contained zeros. These have been removed to enable fitting.')
-        if len(right_censored)!=len(rc0):
+        if len(right_censored) != len(rc0):
             print('WARNING: right_censored contained zeros. These have been removed to enable fitting.')
 
         all_data = np.hstack([failures, right_censored])
@@ -3263,4 +3264,212 @@ class Fit_Beta_2P:
         LL_rc = 0
         LL_f += Fit_Beta_2P.logf(T_f, params[0], params[1]).sum()  # failure times
         LL_rc += Fit_Beta_2P.logR(T_rc, params[0], params[1]).sum()  # right censored times
+        return -(LL_f + LL_rc)
+
+
+class Fit_Loglogistic_2P:
+    '''
+    Fit_Loglogistic_2P
+
+    Fits a 2-parameter Loglogistic distribution (alpha,beta) to the data provided.
+
+    Inputs:
+    failures - an array or list of failure data
+    right_censored - an array or list of right censored data
+    show_probability_plot - True/False. Defaults to True.
+    print_results - True/False. Defaults to True. Prints a dataframe of the point estimate, standard error, Lower CI and Upper CI for each parameter.
+    initial_guess_method - 'scipy' OR 'least squares'. Default is 'least squares'. Both do not take into account censored data but scipy uses MLE, and least squares is least squares regression of the plotting positions. Least squares is generally more accurate.
+    optimizer - 'L-BFGS-B' OR 'TNC'. These are both bound constrained methods. If the bounded method fails, nelder-mead will be used. If nelder-mead fails then the initial guess will be returned with a warning. For more information on optimizers see https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.minimize.html#scipy.optimize.minimize
+    CI - confidence interval for estimating confidence limits on parameters. Must be between 0 and 1. Default is 0.95 for 95% CI.
+    CI_type - time, reliability, None. Default is time. This is the confidence bounds on time or on reliability. Use None to turn off the confidence intervals.
+    kwargs are accepted for the probability plot (eg. linestyle, label, color)
+
+    outputs:
+    success - Whether the solution was found by autograd (True/False)
+        if success is False a warning will be printed indicating that scipy's fit was used as autograd failed. This fit will not be accurate if
+        there is censored data as scipy does not have the ability to fit censored data. Failure of autograd to find the solution should be rare and
+        if it occurs, it is likely that the distribution is an extremely bad fit for the data. Try scaling your data, removing extreme values, or using
+        another distribution.
+    alpha - the fitted Loglogistic_2P alpha parameter
+    beta - the fitted Loglogistic_2P beta parameter
+    loglik - Log Likelihood (as used in Minitab and Reliasoft)
+    loglik2 - Log Likelihood * -2 (as used in JMP Pro)
+    AICc - Akaike Information Criterion corrected
+    BIC - Bayesian Information Criterion
+    distribution - a Loglogistic_Distribution object with the parameters of the fitted distribution
+    alpha_SE - the standard error (sqrt(variance)) of the parameter
+    beta_SE - the standard error (sqrt(variance)) of the parameter
+    Cov_alpha_beta - the covariance between the parameters
+    alpha_upper - the upper CI estimate of the parameter
+    alpha_lower - the lower CI estimate of the parameter
+    beta_upper - the upper CI estimate of the parameter
+    beta_lower - the lower CI estimate of the parameter
+    results - a dataframe of the results (point estimate, standard error, Lower CI and Upper CI for each parameter)
+    '''
+
+    def __init__(self, failures=None, right_censored=None, show_probability_plot=True, print_results=True, CI=0.95, CI_type='time', initial_guess_method='least squares', optimizer='L-BFGS-B', **kwargs):
+        if failures is None or len(failures) < 2:
+            raise ValueError('Maximum likelihood estimates could not be calculated for these data. There must be at least two failures to calculate Loglogistic parameters.')
+
+        if CI <= 0 or CI >= 1:
+            raise ValueError('CI must be between 0 and 1. Default is 0.95 for 95% Confidence interval.')
+
+        if initial_guess_method in ['scipy', 'Scipy', 'SP', 'sp']:
+            initial_guess_method = 'scipy'
+        elif initial_guess_method in ['least_squares', 'least squares', 'Least_squares', 'Least squares', 'Least Squares', 'Least_Squares', 'ls', 'LS']:
+            initial_guess_method = 'least squares'
+        else:
+            raise ValueError('initial_guess_method must be either "scipy" or "least squares". Default is "least squares".')
+
+        if optimizer not in ['L-BFGS-B', 'TNC']:
+            raise ValueError('optimizer must be either "L-BFGS-B" OR "TNC". Default is "L-BFGS-B".')
+
+        # fill with empty lists if not specified
+        if right_censored is None:
+            right_censored = []
+
+        # adjust inputs to be arrays
+        if type(failures) == list:
+            failures = np.array(failures)
+        if type(failures) != np.ndarray:
+            raise TypeError('failures must be a list or array of failure data')
+        if type(right_censored) == list:
+            right_censored = np.array(right_censored)
+        if type(right_censored) != np.ndarray:
+            raise TypeError('right_censored must be a list or array of right censored failure data')
+
+        # remove zeros. These are impossible since the pdf should be 0 at t=0. Leaving them in causes an error.
+        rc0 = right_censored
+        f0 = failures
+        right_censored = rc0[rc0 != 0]
+        failures = f0[f0 != 0]
+        if len(failures) != len(f0):
+            print('WARNING: failures contained zeros. These have been removed to enable fitting.')
+        if len(right_censored) != len(rc0):
+            print('WARNING: right_censored contained zeros. These have been removed to enable fitting.')
+
+        all_data = np.hstack([failures, right_censored])
+        if min(all_data) < 0:
+            raise ValueError('All failure and censoring times must be greater than zero.')
+        self.gamma = 0
+        # Obtain initial guess using either Least squares or Scipy
+        if initial_guess_method == 'least squares':
+            # obtain least squares estimate based on the plotting positions
+            from reliability.Probability_plotting import plotting_positions
+            x, y = plotting_positions(failures=failures, right_censored=right_censored)
+            x_linearised = np.log(x)
+            y_linearised = np.log(1 / np.array(y) - 1)
+            slope, intercept, _, _, _ = ss.linregress(x_linearised, y_linearised)
+            LS_beta = -slope
+            LS_alpha = np.exp(intercept / LS_beta)
+            guess = [LS_alpha, LS_beta]
+        else:  # scipy method of obtaining the initial guess
+            all_data = np.hstack([failures, right_censored])
+            sp = ss.fisk.fit(all_data, floc=0, optimizer='powell')  # scipy's answer is used as an initial guess. Scipy is only correct when there is no censored data
+            if sp[0] > 100:  # correction to high beta which generally occurs with heavily censored datasets
+                guess = [sp[2] * 1.5, sp[0] * 0.4]
+            else:
+                guess = [sp[2], sp[0]]
+        self.initial_guess = guess
+
+        warnings.filterwarnings('ignore')  # it is necessary to supress the warning about the jacobian when using the nelder-mead optimizer
+        n = len(all_data)
+        delta_BIC = 1
+        BIC_array = [1000000]
+        runs = 0
+        bnds = [(0, None), (0, None)]  # bounds on the solution. Helps a lot with stability
+        k = len(guess)
+        while delta_BIC > 0.001 and runs < 5:  # exits after BIC convergence or 5 iterations
+            runs += 1
+            result = minimize(value_and_grad(Fit_Loglogistic_2P.LL), guess, args=(failures, right_censored), jac=True, method=optimizer, bounds=bnds)
+            params = result.x
+            guess = [params[0], params[1]]
+            LL2 = 2 * Fit_Loglogistic_2P.LL(guess, failures, right_censored)
+            BIC_array.append(np.log(n) * k + LL2)
+            delta_BIC = abs(BIC_array[-1] - BIC_array[-2])
+
+        if result.success is True:
+            params = result.x
+            self.success = True
+            self.alpha = params[0]
+            self.beta = params[1]
+        else:  # if the L-BFGS-B optimizer fails then we have a second attempt using the slower but slightly more reliable nelder-mead optimizer
+            guess = self.initial_guess
+            result = minimize(value_and_grad(Fit_Loglogistic_2P.LL), guess, args=(failures, right_censored), jac=True, tol=1e-4, method='nelder-mead')
+            if result.success is True:
+                params = result.x
+                self.success = True
+                self.alpha = params[0]
+                self.beta = params[1]
+            else:
+                self.success = False
+                print('WARNING: Fitting using Autograd FAILED for Loglogistic_2P. The fit from Scipy was used instead so results may not be accurate.')
+                self.alpha = self.initial_guess[0]
+                self.beta = self.initial_guess[1]
+
+        params = [self.alpha, self.beta]
+        LL2 = 2 * Fit_Loglogistic_2P.LL(params, failures, right_censored)
+        self.loglik2 = LL2
+        self.loglik = LL2 * -0.5
+        if n - k - 1 > 0:
+            self.AICc = 2 * k + LL2 + (2 * k ** 2 + 2 * k) / (n - k - 1)
+        else:
+            self.AICc = 'Insufficient data'
+        self.BIC = np.log(n) * k + LL2
+
+        # confidence interval estimates of parameters
+        Z = -ss.norm.ppf((1 - CI) / 2)
+        hessian_matrix = hessian(Fit_Loglogistic_2P.LL)(np.array(tuple(params)), np.array(tuple(failures)), np.array(tuple(right_censored)))
+        covariance_matrix = np.linalg.inv(hessian_matrix)
+        self.alpha_SE = abs(covariance_matrix[0][0]) ** 0.5
+        self.beta_SE = abs(covariance_matrix[1][1]) ** 0.5
+        self.Cov_alpha_beta = abs(covariance_matrix[0][1])
+        self.alpha_upper = self.alpha * (np.exp(Z * (self.alpha_SE / self.alpha)))
+        self.alpha_lower = self.alpha * (np.exp(-Z * (self.alpha_SE / self.alpha)))
+        self.beta_upper = self.beta * (np.exp(Z * (self.beta_SE / self.beta)))
+        self.beta_lower = self.beta * (np.exp(-Z * (self.beta_SE / self.beta)))
+
+        Data = {'Parameter': ['Alpha', 'Beta'],
+                'Point Estimate': [self.alpha, self.beta],
+                'Standard Error': [self.alpha_SE, self.beta_SE],
+                'Lower CI': [self.alpha_lower, self.beta_lower],
+                'Upper CI': [self.alpha_upper, self.beta_upper]}
+        df = pd.DataFrame(Data, columns=['Parameter', 'Point Estimate', 'Standard Error', 'Lower CI', 'Upper CI'])
+        self.results = df.set_index('Parameter')
+        self.distribution = Loglogistic_Distribution(alpha=self.alpha, beta=self.beta, alpha_SE=self.alpha_SE, beta_SE=self.beta_SE, Cov_alpha_beta=self.Cov_alpha_beta, CI=CI, CI_type=CI_type)
+
+        if print_results is True:
+            pd.set_option('display.width', 200)  # prevents wrapping after default 80 characters
+            pd.set_option('display.max_columns', 9)  # shows the dataframe without ... truncation
+            if CI * 100 % 1 == 0:
+                CI_rounded = int(CI * 100)
+            else:
+                CI_rounded = CI * 100
+            print(str('Results from Fit_Loglogistic_2P (' + str(CI_rounded) + '% CI):'))
+            print(self.results)
+            print('Log-Likelihood:', self.loglik, '\n')
+
+        if show_probability_plot is True:
+            pass
+            # from reliability.Probability_plotting import Loglogistic_probability_plot
+            # if len(right_censored) == 0:
+            #     rc = None
+            # else:
+            #     rc = right_censored
+            # Loglogistic_probability_plot(failures=failures, right_censored=rc, __fitted_dist_params=self, CI=CI, CI_type=CI_type, **kwargs)
+
+    @staticmethod
+    def logf(t, a, b):  # Log PDF (2 parameter Loglogistic)
+        return anp.log(b / a) - (b + 1) * anp.log(t / a) - 2 * anp.log(1 + (t / a) ** -b)
+
+    @staticmethod
+    def logR(t, a, b):  # Log SF (2 parameter Loglogistic)
+        return -anp.log((1 + (t / a) ** (b)))
+
+    @staticmethod
+    def LL(params, T_f, T_rc):  # log likelihood function (2 parameter Loglogistic)
+        LL_f = 0
+        LL_rc = 0
+        LL_f += Fit_Loglogistic_2P.logf(T_f, params[0], params[1]).sum()  # failure times
+        LL_rc += Fit_Loglogistic_2P.logR(T_rc, params[0], params[1]).sum()  # right censored times
         return -(LL_f + LL_rc)